Bail out of reducing depency trees on huge dependency conflict sets
This is a simple fix for #6114, in which sometimes conflict sets get so huge that
the current reducing code (which is based on Array#combination) breaks down.

By simply bailing out when the conflict set is bigger than a certain size,
we'll get a result that is similar to what happened in earlier bundler versions
but skip a ton of CPU cycles and memory allocations.

I've chosen the limit rather unscientifically by playing around with the result
set sizes. 15 seems to be a good compromise but really anything larger than 10
should work (and with work I mean "should not usually not have to be invoked").