Be more precise when replacing Bundler version in lock file.
Lets say the current Bundler version is 1.16.0. If the test suite is, by
a chance, executed in directory which contains such version string, this
path is stored in Gemfile.lock file. Later, the test tries to replace
the version of Bundler in the lock file, but instead, it replaces the
version in path. Be more careful what should be actually replaced.

This is related to #6185 although it does not resolve the original
concern.