Fix failing spec in the bundle exec system gem spec
There was an issue where a spec was failing because it was activating
the version of bundler inside of RubyGems and not the development
version that we're working on.

This would cause the Lockfile to raise an error because we were generating
a Bundler 2 lockfile but Bundler 1 was being activated.