Bump the bundle_binstub check-length to 300 characters
Some packaging systems require modifications to the hash-bang
interpreter path, making them longer than normal. One such example of
this is the habitat-sh/habitat project. The result is that the 150
character limit does not find the regex it is looking for, and prevents
an otherwise valid binary from starting.

This change doubles the length of the check from 150 characters to 300
characters. This change has been validated in an impacted environment.

Signed-off-by: Tom Duffield <tom@chef.io>