Make all plugin events registered to make documenting them easier
Every event in #hook and #add_hook will check if the event is registered in
Bundler::Plugin::Events. This allows for easy tracking of what's calling events,
and allow documentation to easily point to a single location.

It also makes testing easier as events are predicatable and accessible via constants