Fix DepProxy#hash calculation
According to the official Ruby documentation, "the eql? method returns true if obj and other
refer to the same hash key." This was not the case as the hash key of a DepProxy instance was
only calculated based on the dep object but in the eql? method dep and platform attributes were
computed. This caused that equal objects returned different hash keys.

https://ruby-doc.org/core-2.5.1/Object.html#method-i-eql-3F