Lower required RubyGems to 2.5 or greater
It turns out that every existing version of RubyGems will ignore the
required RubyGems version and try to install the newest version of
Bundler... until the middle of the install, when it will raise an
exception and fail.

There are some bugfixes in RubyGems 3 that improve the quality of life
for Bundler 2 users, but they are less bad than RubyGems simply
exploding anytime you try to run `gem install bundler`.