Move Rubocop development dependency into gemspec
Now that Bundler no longer supports Ruby <2.3, the Ruby minimum version
requirement that prevented Rubocop from being specified as a development
dependency in the gemspec no longer applies.

Having Rubocop in the gemspec is useful for tools like RubyGems.org or
Bundix that use it as the source of truth for dependencies.