Stop leaking to system rubygems on subprocesses
In order to be able to test against different versions of rubygems, we
prepend our local copy of rubygems to the $LOAD_PATH. This is done as
early as possible, inside each binstubs through [this helper file].

However, some subprocesses overwrite the `RUBYOPT` env variable, thus
leaking to the system copy of rubygems again. Since we already [prepend
the hacks file] to `RUBYOPT` in the main helper, while also preserving
the previous value, this "customized environments" are not needed.