Refactor list specs to not change gemfile sources
Previously, all specs would bundle a Gemfile from the "repo1" source by
default. Then, some list specs would end up using a "repo2" source. The
behavior when changing repo sources is changing with the
`disable_multisource` setting, but it is unrelated to the list command.

So, I refactored the list command specs to not change sources, and
extracted the behavior change about changing sources to a separate spec.