Previous approach for not walking further up from test folder
I tried monkeypatching the gemfile finder methods in hax.rb, but that
resulted in loading parts of bundler too early, and getting redefinition
warnings when loading them again from exe/bundle, resulting in actual
errors in some cases. The fix for that required eagerly loading the
monkeypatched parts of bundler, but that involves changing lib just for
the sake of making existing tests pass, and that's what the original
changes intend to avoid.

So, I'm just reverting to the previous approach of setting an
environment variable just for testing and check that.

Reverts 02028261def98051e0aec14191524f15d64b6669.