Change the refresh spec to not use `stubs_for`
We plan to incrementally populate this array in rubygems, so that will
break the test because the first time `find_by_name("rack")` is called
will make the `rack` gem available in the array.

We can test the same thing using some other way, which is more
appropriate anyways, because `Bundler.rubygems.find_name` is only used
inside `bundler` to look for the `bundler` gem itself whereas
`Bundler.rubygems.all_specs` is used in more places.