Ignore whitespace when checking docs
Different versions of `groff` seem to generate slight differences in
whitespace. It is too much burden for contributors in my opinion to make
them install the exact groff version CI is running, and whitespace
amount in man pages doesn't seem to change the rendered output anyways,
so it should be good I think.