Remove unnecessary lines from `bundler/setup`
Bundler no longer needs to be in the `$LOAD_PATH` once a version of
itself has been loaded.

In any case, the `Metadata` source takes care of adding `bundler` to the
index:

https://github.com/bundler/bundler/blob/851a12519d55dbed9af1268605e258a01c819cd2/lib/bundler/source/metadata.rb#L13-L30

and then that spec is added to the resolution here:

https://github.com/bundler/bundler/blob/851a12519d55dbed9af1268605e258a01c819cd2/lib/bundler/definition.rb#L180-L184

And from the resulting set of specs, the `$LOAD_PATH` is setup:

https://github.com/bundler/bundler/blob/851a12519d55dbed9af1268605e258a01c819cd2/lib/bundler/runtime.rb#L25-L38

So `bundler` will be present in the `$LOAD_PATH` anyways, and the lines
being removed here will never be useful.