Split default gem activation specs
Currently they're using a single Gemfile with all default gems, and
making sure that all default gems can be activated. It a single default
gem is activated by bundler, this results in huge errors messages
because the spec still tries to activate all default gems present in the
Gemfile, and fails for all of them.

Instead, we separate on spec for each default gem, which makes error
messages smaller, and helps figuring out the root cause of the issue. In
this case, the `uri` gem.