Improve platform specific gem resolution
It resolves #6247. This changes includes the patches that add (RSpec)
specs for this situation in #6247.

If there is a platform specific gem but it can't be resolved available
version, Bundler reports an error.

For example,

    @index = build_index do
      gem "bar", "1.0.0"
      gem "foo", "1.0.0"
      gem "foo", "1.0.0", "x64-mingw32" do
        dep "bar", "< 1"
      end
    end
    dep "foo"
    platforms "x64-mingw32"

raises an error because foo-1.0.0-x64-mingw32 requires bar<1 but there
isn't bar<1.

With this change, foo-1.0.0 (no x64-mingw32) is used as
fallback. Because foo-1.0.0 doesn't depend on bar<1.